requirements.txt
# CORE DATABASE & ENVIRONMENT
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# LANGCHAIN / LANGGRAPH / LANGSMITH
langchain==0.3.24
langchain-core>=0.3.60,<0.4.0
langchain-community>=0.3.23,<0.4.0
langchain-openai>=0.3.14,<0.4.0
langgraph==0.4.0
langsmith==0.3.45
langfuse

# VECTOR STORES / EMBEDDINGS
chromadb>=1.0.9
langchain-chroma>=0.2.6,<0.3.0

# LLM PROVIDERS
openai==1.76.0
anthropic>=0.41.0
langchain-anthropic>=0.3.3
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
cohere==5.13.0
langchain-cohere>=0.4.0


# SERVER & API
fastapi==0.115.4
uvicorn==0.30.6

# AUXILIARIES / UTILITIES
tqdm==4.67.1
requests==2.32.3
numpy==2.2.5
pandas==2.3.3
pydantic==2.11.3
SQLAlchemy==2.0.40

# INDUSTRY CONTEXT / DATA SOURCES
yfinance==0.2.66
yahooquery==2.4.1
wbdata==1.1.0

# DEV / TESTING / DASHBOARD
pytest==8.3.3
streamlit==1.39.0
